Split legacy media overlays
